code/requirements.txt
altgraph==0.17
bitarray==2.2.5
more-itertools==8.8.0
Pillow==8.3.1
